+++ redback/__init__.py
@@ -2,3 +2,7 @@ from redback import analysis, constants, get_data, redback_errors, priors, resul
     transient_models, utils, photosphere, sed, interaction_processes, constraints, plotting, model_library, simulate_transients
 from redback.transient import afterglow, kilonova, prompt, supernova, tde
 from redback.sampler import fit_model
+from redback.utils import setup_logger
+
+__version__ = "1.0.2"
+setup_logger(log_level='info')