+++ redback/constraints.py
@@ -2,6 +2,7 @@ import numpy as np
 import redback.eos as eos
 from redback.constants import *
 from redback.utils import calc_tfb
+from scipy.interpolate import interp1d
 
 def slsn_constraint(parameters):
     """
@@ -23,7 +24,7 @@ def slsn_constraint(parameters):
     neutrino_energy = 1e51
     total_energy = kinetic_energy + neutrino_energy
     # ensure rotational energy is greater than total output energy
-    converted_parameters['erot_constraint'] = rotational_energy - total_energy
+    converted_parameters['erot_constraint'] = total_energy/rotational_energy
     # ensure t_nebula is greater than 100 days
     converted_parameters['t_nebula_min'] = tnebula - 100
     return converted_parameters
@@ -43,7 +44,7 @@ def basic_magnetar_powered_sn_constraints(parameters):
     kinetic_energy = 0.5 * mej * vej**2
     rotational_energy = 2.6e52 * (mass_ns/1.4)**(3./2.) * p0**(-2)
     # ensure rotational energy is greater than total output energy
-    converted_parameters['erot_constraint'] = rotational_energy - kinetic_energy
+    converted_parameters['erot_constraint'] = kinetic_energy/rotational_energy
     return converted_parameters
 
 def general_magnetar_powered_sn_constraints(parameters):
@@ -61,7 +62,7 @@ def general_magnetar_powered_sn_constraints(parameters):
     tau = parameters['tsd']
     rotational_energy = 2*l0*tau
     # ensure rotational energy is greater than total output energy
-    converted_parameters['erot_constraint'] = rotational_energy - kinetic_energy
+    converted_parameters['erot_constraint'] = kinetic_energy/rotational_energy
     return converted_parameters
     
 def general_magnetar_powered_supernova_constraints(parameters):
@@ -77,7 +78,7 @@ def general_magnetar_powered_supernova_constraints(parameters):
     nn = parameters['nn']    
     rotational_energy = (nn-1)*l0*tau/2.0
     # ensure rotational energy is less than the maximum spin down energy
-    converted_parameters['erot_constraint'] = 1e53 - rotational_energy
+    converted_parameters['erot_constraint'] = rotational_energy/1e53
     return converted_parameters    
 
 def tde_constraints(parameters):
@@ -91,7 +92,7 @@ def tde_constraints(parameters):
     rp = parameters['pericenter_radius']
     mass_bh = parameters['mass_bh']
     schwarzchild_radius = (2 * graviational_constant * mass_bh * solar_mass /(speed_of_light**2))/au_cgs
-    converted_parameters['disruption_radius'] = rp - schwarzchild_radius
+    converted_parameters['disruption_radius'] = schwarzchild_radius/rp
     return converted_parameters
 
 def gaussianrise_tde_constraints(parameters):
@@ -103,13 +104,11 @@ def gaussianrise_tde_constraints(parameters):
     converted_parameters = parameters.copy()
     ms = parameters['stellar_mass']
     mbh6 = parameters['mbh_6']
-    etamin = 0.01*(ms**(-7./15.))*(mbh6**(2./3.))
     betamax = 12.*(ms**(7./15.))*(mbh6**(-2./3.))
     tfb = calc_tfb(binding_energy_const=0.8, mbh_6=mbh6,stellar_mass=ms)/86400
     tfb_obs = tfb * (1 + parameters['redshift'])
-    converted_parameters['eta_low'] = converted_parameters['eta'] - etamin
-    converted_parameters['beta_high'] = betamax - converted_parameters['beta']
-    converted_parameters['tfb_max'] = tfb_obs - converted_parameters['peak_time']
+    converted_parameters['beta_high'] = converted_parameters['beta']/betamax
+    converted_parameters['tfb_max'] = converted_parameters['peak_time']/tfb_obs
     return converted_parameters
 
 def nuclear_burning_constraints(parameters):
@@ -126,7 +125,7 @@ def nuclear_burning_constraints(parameters):
     kinetic_energy = 0.5 * mej * (vej / 2.0) ** 2
     excess_constant = -(56.0 / 4.0 * 2.4249 - 53.9037) / proton_mass * mev_cgs
     emax = excess_constant * mej * fnickel
-    converted_parameters['emax_constraint'] = emax - kinetic_energy
+    converted_parameters['emax_constraint'] = kinetic_energy/emax
     return converted_parameters
 
 def simple_fallback_constraints(parameters):
@@ -149,7 +148,7 @@ def simple_fallback_constraints(parameters):
     neutrino_energy = 1e51
     total_energy = e_fallback + neutrino_energy
     # ensure total energy is greater than kinetic energy
-    converted_parameters['en_constraint'] = total_energy - kinetic_energy
+    converted_parameters['en_constraint'] = kinetic_energy/total_energy
     # ensure t_nebula is greater than 100 days
     converted_parameters['t_nebula_min'] = tnebula - 100
     return converted_parameters
@@ -162,15 +161,18 @@ def csm_constraints(parameters):
     :param parameters: dictionary of parameters
     :return: converted_parameters dictionary where the violated samples are thrown out
     """
-    from redback.utils import get_csm_properties
     converted_parameters = parameters.copy()
     mej = parameters['mej']
     csm_mass = parameters['csm_mass']
     kappa = parameters['kappa']
     r0 = parameters['r0']
     vej = parameters['vej']
-    nn = parameters.get('nn', np.ones(len(mej)) * 12.)
-    delta = parameters.get('delta', np.ones(len(mej)))
+    if hasattr(parameters['mej'], "__len__"):
+        nn = parameters.get('nn', np.ones(len(mej)) * 8.)
+        delta = parameters.get('delta', np.ones(len(mej)))
+    else:
+        nn = parameters.get('nn', 12.)
+        delta = parameters.get('delta', 0.)
     eta = parameters['eta']
     rho = parameters['rho']
 
@@ -180,12 +182,16 @@ def csm_constraints(parameters):
     vej = vej * km_cgs
     Esn = 3. * vej ** 2 * mej / 10.
 
-    AA = np.zeros(len(mej))
-    Bf = np.zeros(len(mej))
-    for x in range(len(mej)):
-        csm_properties = get_csm_properties(nn[x], eta[x])
-        AA[x] = csm_properties.AA
-        Bf[x] = csm_properties.Bf
+    ns = [6, 7, 8, 9, 10, 12, 14]
+    Bfs = [1.377, 1.299, 1.267, 1.250, 1.239, 1.226, 1.218]
+    As = [0.62, 0.27, 0.15, 0.096, 0.067, 0.038, 0.025]
+
+    Bf_func = interp1d(ns, Bfs)
+    A_func = interp1d(ns, As)
+
+    Bf = Bf_func(nn)
+    AA = A_func(nn)
+
     qq = rho * r0 ** eta
     # outer CSM shell radius
     radius_csm = ((3.0 - eta) / (4.0 * np.pi * qq) * csm_mass + r0 ** (3.0 - eta)) ** (
@@ -203,14 +209,14 @@ def csm_constraints(parameters):
                    (3.0 - delta) * (nn - 3.0) * mej) ** ((nn - 5.0) / 2.0))
 
     tshock = ((radius_csm - r0) / Bf / (AA * g_n / qq) ** (
-                       1. / (nn - eta))) ** ((nn - eta) /(nn - 3))
+                       1. / (nn - eta))) ** ((nn - eta) / (nn - 3))
 
-    diffusion_time = np.sqrt(2. * kappa * mass_csm_threshold /(vej * 13.7 * 3.e10))
+    diffusion_time = np.sqrt(2. * kappa * mass_csm_threshold / (vej * 13.7 * 3.e10))
     # ensure shock crossing time is greater than diffusion time
-    # converted_parameters['shock_time'] = tshock - diffusion_time
+    converted_parameters['shock_time'] = diffusion_time/tshock
     # ensure photospheric radius is within the csm i.e., r_photo < radius_csm and r_photo > r0
-    converted_parameters['photosphere_constraint_1'] = radius_csm - r_photosphere
-    converted_parameters['photosphere_constraint_2'] = r_photosphere - r0
+    converted_parameters['photosphere_constraint_1'] = r_photosphere/radius_csm
+    converted_parameters['photosphere_constraint_2'] = r0/r_photosphere
     return converted_parameters
 
 def piecewise_polytrope_eos_constraints(parameters):