+++ redback/transient/afterglow.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import os
 from os.path import join
 from typing import Union
+import re
 
 import numpy as np
 import pandas as pd
@@ -57,7 +58,7 @@ class Afterglow(Transient):
         :type flux: np.ndarray, optional
         :type flux_err: np.ndarray, optional
         :param flux_err: Flux error values.
-        :param flux_density:Flux density values.
+        :param flux_density: Flux density values.
         :type flux_density: np.ndarray, optional
         :param flux_density_err: Flux density error values.
         :type flux_density_err: np.ndarray, optional
@@ -244,7 +245,8 @@ class Afterglow(Transient):
                 'BAT Photon Index (15-150 keV) (PL = simple power-law, CPL = cutoff power-law)'].fillna(0)
             self.meta_data = meta_data
         except FileNotFoundError:
-            logger.warning("Meta data does not exist for this event.")
+            logger.info("Metadata does not exist for this event.")
+            logger.info("Setting metadata to None. This is not an error, but a warning that no metadata could be found online.")
             self.meta_data = None
 
     def _set_photon_index(self) -> None:
@@ -305,9 +307,13 @@ class Afterglow(Transient):
         :return: The cleaned string converted into a float.
         :rtype: float
         """
-        for r in ["PL", "CPL", ",", "C", "~", " ", 'Gemini:emission', '()']:
-            string = string.replace(r, "")
-        return float(string)
+        try:
+            for r in ["PL", "CPL", ",", "C", "~", " ", 'Gemini:emission', '()']:
+                string = string.replace(r, "")
+                new_float = float(string)
+        except ValueError:
+            new_float = float(re.findall("\d+\.\d+", string)[0])
+        return new_float
 
     def analytical_flux_to_luminosity(self) -> None:
         """Converts flux to luminosity using the analytical method."""