+++ redback/transient_models/combined_models.py
@@ -1,5 +1,11 @@
-import redback.transient_models.extinction_models
-def tophat_and_twolayerstratified(time, redshift, av, thv, loge0, thc, logn0, p, logepse, logepsb, ksin, g0, mej, vej_1, vej_2, kappa, beta, **kwargs):
+import redback.transient_models.extinction_models as em
+import redback.transient_models as tm
+from redback.utils import nu_to_lambda
+from redback.utils import citation_wrapper
+
+@citation_wrapper('https://ui.adsabs.harvard.edu/abs/2020ApJ...896..166R/abstract, https://ui.adsabs.harvard.edu/abs/2020ApJ...891..152H/abstract')
+def tophat_and_twolayerstratified(time, redshift, av, thv, loge0, thc, logn0, p, logepse,
+                                  logepsb, ksin, g0, mej, vej_1, vej_2, kappa, beta, **kwargs):
     
     """
     function to combine the flux density signals of a tophat afterglow and a two layer stratified kilonova with extinction
@@ -34,20 +40,22 @@ def tophat_and_twolayerstratified(time, redshift, av, thv, loge0, thc, logn0, p,
     :return: flux density signal with extinction added
     
     """
-    kwargs['output_format']='flux_density'
-    kwargs['base_model']='tophat'
-    afterglow = redback.transient_models.extinction_models.extinction_with_afterglow_base_model(time=time, redshift=redshift, av=av,
-         thv=thv, loge0=loge0 , thc= thc, logn0=logn0, p=p, logepse=logepse, logepsb=logepsb, ksin=ksin, g0=g0,
-        **kwargs)
-    kwargs['base_model']='two_layer_stratified_kilonova'
-    kilonova = redback.transient_models.extinction_models.extinction_with_kilonova_base_model(time=time, redshift=redshift, av=av,
-          mej=mej, vej_1=vej_1, vej_2=vej_2, kappa=kappa, beta=beta, **kwargs)
-    
+    kwargs['output_format'] = 'flux_density'
+    afterglow = tm.afterglow_models.tophat(time=time, redshift=redshift, thv=thv, loge0=loge0, thc=thc, logn0=logn0,
+                                           p=p, logepse=logepse, logepsb=logepsb, ksin=ksin, g0=g0, **kwargs)
+    kilonova = tm.kilonova_models.two_layer_stratified_kilonova(time=time, redshift=redshift, mej=mej, vej_1=vej_1,
+                                                                vej_2=vej_2, kappa=kappa, beta=beta, **kwargs)
     combined = afterglow+kilonova
+    r_v = kwargs.get('r_v', 3.1)
+    # correct for extinction
+    angstroms = nu_to_lambda(kwargs['frequency'])
+    combined = em._perform_extinction(flux_density=combined, angstroms=angstroms, av=av, r_v=r_v)
     return combined
 
-
-def tophat_and_twocomponent(time, redshift, av, thv, loge0, thc, logn0, p, logepse, logepsb, ksin, g0, mej_1, vej_1, temperature_floor_1, kappa_1, mej_2, vej_2, temperature_floor_2, kappa_2, **kwargs):
+@citation_wrapper('https://ui.adsabs.harvard.edu/abs/2020ApJ...896..166R/abstract, redback')
+def tophat_and_twocomponent(time, redshift, av, thv, loge0, thc, logn0,
+                            p, logepse, logepsb, ksin, g0, mej_1, vej_1,
+                            temperature_floor_1, kappa_1, mej_2, vej_2, temperature_floor_2, kappa_2, **kwargs):
     
     """
     function to combine the flux density signals of a tophat afterglow and a two component kilonova with extinction added
@@ -85,19 +93,22 @@ def tophat_and_twocomponent(time, redshift, av, thv, loge0, thc, logn0, p, logep
     :return: flux density signal with extinction added
     """
     
-    kwargs['output_format']='flux_density'
-    kwargs['base_model']='tophat'
-    afterglow = redback.transient_models.extinction_models.extinction_with_afterglow_base_model(time=time, redshift=redshift, av=av,
-          thv=thv, loge0=loge0 , thc= thc, logn0=logn0, p=p, logepse=logepse, logepsb=logepsb, ksin=ksin, g0=g0,
-        **kwargs)
-    kwargs['base_model']='two_component_kilonova_model'
-    kilonova = redback.transient_models.extinction_models.extinction_with_kilonova_base_model(time=time, redshift=redshift, av=av,
-          mej_1=mej_1, vej_1=vej_2, temperature_floor_1=temperature_floor_1, kappa_1=kappa_1,             mej_2=mej_2, vej_2=vej_2, temperature_floor_2=temperature_floor_2, kappa_2=kappa_2, **kwargs)
+    kwargs['output_format'] = 'flux_density'
+    afterglow = tm.afterglow_models.tophat(time=time, redshift=redshift, thv=thv, loge0=loge0, thc=thc, logn0=logn0,
+                                           p=p, logepse=logepse, logepsb=logepsb, ksin=ksin, g0=g0, **kwargs)
+    kilonova = tm.kilonova_models.two_component_kilonova_model(time=time, redshift=redshift, av=av,
+                                                      mej_1=mej_1, vej_1=vej_1, temperature_floor_1=temperature_floor_1,
+                                                      kappa_1=kappa_1, mej_2=mej_2, vej_2=vej_2,
+                                                      temperature_floor_2=temperature_floor_2, kappa_2=kappa_2, **kwargs)
     
     combined = afterglow + kilonova
+    r_v = kwargs.get('r_v', 3.1)
+    # correct for extinction
+    angstroms = nu_to_lambda(kwargs['frequency'])
+    combined = em._perform_extinction(flux_density=combined, angstroms=angstroms, av=av, r_v=r_v)
     return combined
 
-
+@citation_wrapper('https://ui.adsabs.harvard.edu/abs/2020ApJ...896..166R/abstract, https://ui.adsabs.harvard.edu/abs/1982ApJ...253..785A/abstract')
 def tophat_and_arnett(time, av, redshift, thv, loge0, thc, logn0, p, logepse, logepsb, ksin, g0, f_nickel, mej, **kwargs):
     
     """
@@ -137,46 +148,55 @@ def tophat_and_arnett(time, av, redshift, thv, loge0, thc, logn0, p, logepse, lo
     :return: flux density with extinction added
     """
     
-    kwargs['output_format']='flux_density'
-    kwargs['base_model']='tophat'
-    afterglow = redback.transient_models.extinction_models.extinction_with_afterglow_base_model(time=time, redshift=redshift, av=av,
-         thv=thv, loge0=loge0 , thc= thc, logn0=logn0, p=p, logepse=logepse, logepsb=logepsb, ksin=ksin, g0=g0,
-        **kwargs)
-    kwargs['base_model']='arnett'
-    supernova = redback.transient_models.extinction_models.extinction_with_supernova_base_model(time=time, redshift=redshift, av=av,
-         f_nickel=f_nickel, mej=mej, **kwargs)
-    
+    kwargs['output_format'] = 'flux_density'
+    afterglow = tm.afterglow_models.tophat(time=time, redshift=redshift, thv=thv, loge0=loge0, thc=thc, logn0=logn0,
+                                           p=p, logepse=logepse, logepsb=logepsb, ksin=ksin, g0=g0, **kwargs)
+    kwargs['base_model'] = 'arnett'
+    supernova = tm.supernova_models.arnett(time=time, redshift=redshift, f_nickel=f_nickel, mej=mej, **kwargs)
     combined = afterglow + supernova
+    r_v = kwargs.get('r_v', 3.1)
+    # correct for extinction
+    angstroms = nu_to_lambda(kwargs['frequency'])
+    combined = em._perform_extinction(flux_density=combined, angstroms=angstroms, av=av, r_v=r_v)
     return combined
 
-def afterglow_and_optical(time, redshift, av, model_type, afterglow_kwargs, optical_kwargs, **shared_kwargs):
-    
+@citation_wrapper('redback, and any citations for the specific model you use')
+def afterglow_and_optical(time, redshift, av, **model_kwargs):
     """
     function to combine the signals of any afterglow and any other optical transient with extinction added
     
     :param time: time in days in observer frame
     :param redshift: source redshift
     :param av: absolute mag extinction
-    :param model_type: specify type of optical transient model- 'supernova', 'tde', 'kilonova', 'magnetar_driven', 'shock_powered'
+    :param model_kwargs: kwargs shared by models frequency and r_v (extinction paramater defaults to 3.1)
     :param afterglow_kwargs: dictionary of  parameters required by the afterglow transient model specified by 'base_model'
         and any additional keyword arguments. Refer to model documentation for details.
     :param optical_kwargs: dictionary of parameters required by the optical transient model specifed by 'base_model'
         and any additional keyword arguments. Note the base model must correspond to the given model type. Refer to model documentation
         for details.
-    :param shared_kwargs: kwargs shared by models e.g. output_format, frequency, bands, r_v (extinction paramater defaults to 3.1)
-    :return: set by shared_kwargs output format - 'flux_density' or 'magnitude' with extinction added
-        note that only afterglow_models_sed allow for magnitude outputs
+    :return: flux density signal with extinction added
     """
-    
-    afterglow_kwargs.update(shared_kwargs)
-    optical_kwargs.update(shared_kwargs)
-   
-    afterglow = redback.transient_models.extinction_models._evaluate_extinction_model(time=time, redshift=redshift, av=av,
-                                                                                   model_type='afterglow', **afterglow_kwargs)
-    
-    optical= redback.transient_models.extinction_models._evaluate_extinction_model(time=time, redshift=redshift, av=av,
-                                                                   model_type=model_type, **optical_kwargs)
-  
-    combined= afterglow + optical
+
+    from redback.model_library import all_models_dict
+    optical_kwargs = model_kwargs['optical_kwargs']
+    afterglow_kwargs = model_kwargs['afterglow_kwargs']
+
+    _afterglow_kwargs = afterglow_kwargs.copy()
+    _afterglow_kwargs.update(model_kwargs)
+
+    _optical_kwargs = optical_kwargs.copy()
+    _optical_kwargs.update(model_kwargs)
+
+    afterglow_function = all_models_dict[_afterglow_kwargs['base_model']]
+    afterglow = afterglow_function(time=time, redshift=redshift, output_format='flux_density', **_afterglow_kwargs)
+
+    optical_function = all_models_dict[_optical_kwargs['base_model']]
+    optical = optical_function(time=time, redshift=redshift, output_format='flux_density', **_optical_kwargs)
+
+    combined = afterglow + optical
+    r_v = model_kwargs.get('r_v', 3.1)
+    # correct for extinction
+    angstroms = nu_to_lambda(model_kwargs['frequency'])
+    combined = em._perform_extinction(flux_density=combined, angstroms=angstroms, av=av, r_v=r_v)
     return combined