+++ redback/transient_models/extinction_models.py
@@ -1,6 +1,5 @@
 from inspect import isfunction
 import numpy as np
-
 import redback.utils
 from redback.transient_models.fireball_models import predeceleration
 from redback.utils import logger, calc_ABmag_from_flux_density, citation_wrapper, lambda_to_nu
@@ -12,7 +11,14 @@ extinction_afterglow_base_models = ['tophat', 'cocoon', 'gaussian',
                                     'kn_afterglow', 'cone_afterglow',
                                     'gaussiancore', 'gaussian',
                                     'smoothpowerlaw', 'powerlawcore',
-                                    'tophat']
+                                    'tophat','tophat_from_emulator',
+                                    'kilonova_afterglow_redback', 'kilonova_afterglow_nakarpiran',
+                                    'tophat_redback', 'gaussian_redback', 'twocomponent_redback',
+                                    'powerlaw_redback', 'alternativepowerlaw_redback', 'doublegaussian_redback',
+                                    'tophat_redback_refreshed', 'gaussian_redback_refreshed',
+                                    'twocomponent_redback_refreshed','powerlaw_redback_refreshed',
+                                    'alternativepowerlaw_redback_refreshed', 'doublegaussian_redback_refreshed']
+
 extinction_integrated_flux_afterglow_models = extinction_afterglow_base_models
 extinction_supernova_base_models = ['sn_exponential_powerlaw', 'arnett', 'shock_cooling_and_arnett',
                                     'basic_magnetar_powered', 'slsn', 'magnetar_nickel',
@@ -20,16 +26,17 @@ extinction_supernova_base_models = ['sn_exponential_powerlaw', 'arnett', 'shock_
                                     'general_magnetar_slsn','general_magnetar_driven_supernova']
 extinction_kilonova_base_models = ['nicholl_bns', 'mosfit_rprocess', 'mosfit_kilonova',
                                    'power_law_stratified_kilonova','bulla_bns_kilonova',
-                                   'bulla_nsbh_kilonova', 'kasen_nsbh_kilonova','two_layer_stratified_kilonova',
+                                   'bulla_nsbh_kilonova', 'kasen_bns_kilonova','two_layer_stratified_kilonova',
                                    'three_component_kilonova_model', 'two_component_kilonova_model',
                                    'one_component_kilonova_model', 'one_component_ejecta_relation',
                                    'one_component_ejecta_relation_projection', 'two_component_bns_ejecta_relation',
                                    'polytrope_eos_two_component_bns', 'one_component_nsbh_ejecta_relation',
                                    'two_component_nsbh_ejecta_relation','metzger_kilonova_model']
-extinction_tde_base_models = ['tde_analytical', 'tde_semianalytical', 'gaussianrise_metzger_tde']
+extinction_tde_base_models = ['tde_analytical', 'tde_semianalytical', 'gaussianrise_cooling_envelope',
+                              'cooling_envelope', 'bpl_cooling_envelope']
 extinction_magnetar_driven_base_models = ['basic_mergernova', 'general_mergernova', 'general_mergernova_thermalisation',
                                           'general_mergernova_evolution', 'metzger_magnetar_driven_kilonova_model',
-                                          'general_metzger_magnetar_driven', 'general_magnetar_driven_thermalisation',
+                                          'general_metzger_magnetar_driven', 'general_metzger_magnetar_driven_thermalisation',
                                           'general_metzger_magnetar_driven_evolution']
 extinction_shock_powered_base_models = ['shocked_cocoon', 'shock_cooling']
 
@@ -80,7 +87,6 @@ def _perform_extinction(flux_density, angstroms, av, r_v):
     :return: flux
     """
     import extinction  # noqa
-    import numpy.ma as ma
     if isinstance(angstroms, float):
         angstroms = np.array([angstroms])    
     mag_extinction = extinction.fitzpatrick99(angstroms, av, r_v=r_v)