+++ CHANGELOG.md
@@ -1,5 +1,53 @@
 # All notable changes will be documented in this file
 
+## [1.0.2] 2024-04-18
+Version 1.0.2 release of redback
+
+### Added
+- Tophat_redback surrogate model
+- New SNCosmo fitting and lightcurve generation example
+- New ToO simulations example 
+- New heating rate for kilonova following Rosswog and Korobkin 2022c
+- New kilonova models using Rosswog and Korobkin 2022c heating rate prescription 
+- Fractional error likelihood. 
+- Phenomenological model for emission/absorption lines in spectrum.
+
+### Changed
+- Changed inbuilt constraints to be easier for end users to define in priors
+- Logger now outputs info as standard
+- Maximum likelihood estimation for all likelihoods
+- Bugfix for SNCosmo source when evaluating magnitudes with spectrum that go to zero. 
+- Update simulation interface for changing noise types.  
+- Fix CSM prior and logic for forward and reverse shock masks. 
+- LASAIR API
+
+## [1.0.0] 2023-08-25
+Version 1.0.1 release of redback
+
+### Changed
+- Plotting behaviour for band scaling with magnitudes.
+- Added reference and citation to redback paper now on arXiv to documentation/tutorials/readme
+
+## [1.0.0] 2023-08-24
+Version 1.0.0 release of redback
+
+### Added
+- Simulation interface for generic transients or for Surveys such as ZTF and LSST
+- New TDE model (cooling envelope)
+- New magnetar driven model for SLSN/Ic-BL/Fbot
+- ~15 new afterglow models 
+- New kilonova afterglow models
+- Joint kilonova + afterglow models
+- Added ability to add filters to redback and SNcosmo
+- Plotting spectrum tools
+- Several new examples
+- Several new plotting options e.g., band scaling
+- Redback paper is submitted!
+
+### Changed
+- Plotting max likelihood and random sample colour to a 
+different colour to the data now requires an additional keyword argument 
+
 ## [0.5.0] 2023-04-14
 Version 0.5.0 release of redback